feat: move shared assets to core assets
