+++ src/core/assets/CancelChatIcon.tsx
