+++ src/core/assets/KavaAILogo.tsx
