+++ src/core/assets/LaunchIcon.tsx
