+++ src/core/assets/ResetChatIcon.tsx
