+++ src/core/assets/SendChatIcon.tsx
