+++ src/core/assets/index.ts
