+++ src/core/components/NavBar.tsx
@@ -1,8 +1,8 @@
 import { JSX, useState } from 'react';
 import styles from './NavBar.module.css';
 import DeepseekIcon from '../../features/reasoning/assets/DeepseekIcon';
-import KavaAILogo from '../../shared/assets/KavaAILogo';
-import HamburgerIcon from '../../shared/assets/HamburgerIcon';
+import KavaAILogo from '../assets/KavaAILogo';
+import HamburgerIcon from '../assets/HamburgerIcon';
 import { isInIframe } from '../utils/isInIframe';
 
 const FEAT_UPDATED_DESIGN = import.meta.env.VITE_FEAT_UPDATED_DESIGN;
